robinho/requirements.txt
pandas
scikit-learn>=0.19
imblearn
tornado
jupyter
flake8
marisa-trie
lime
nltk
python-telegram-bot
psutil
requests
chatterbot
python-Levenshtein